V9.3.0 - Support for Friends+Enemies+Targets page
Fixed Profile page that was broken because Torn removed the convenient "empty-block" div that multiple scripts (including BSP) was using.

Support for Friends+Enemy pages is back online (Torn improved the UX for them and it was broken)

Support for the new Targets page.

Added a first implementation of BSP display in the attack page. Disabled by default (go in BSP Settings / Pages to enable it)